Implemented map validation functions, need to check border after
Added code to validate the content and size of the game map in the functions check_map_content() and size_of_map(). The check_map() function now uses these functions to validate the map before proceeding. The function get_map() now displays the map for testing purposes, Further work will be done to validate the border of the map.